backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
marshmallow==3.21.3
pymongo==4.7.0
